requirements.txt
streamlit == 1.32.2
streamlit-extras == 0.4.0
black
mypy
pandas-stubs
pycountry
plotly
email-validator